requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# LLM Integrations
openai==1.10.0
anthropic==0.9.0
google-generativeai==0.3.2
langchain==0.1.4
langchain-openai==0.0.5

# Vector embeddings
sentence-transformers==2.3.1
pgvector==0.2.4

# File handling
python-magic==0.4.27
pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0

# Background tasks
celery==5.3.6
flower==2.0.1

# Utils
httpx==0.26.0
aiofiles==23.2.1
python-slugify==8.0.2
email-validator==2.1.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
